+++ src/core/ObservableQuery.ts
@@ -27,6 +27,8 @@ import {
 import { tryFunctionOrLogError } from '../util/errorHandling';
 
 import { isEqual } from '../util/isEqual';
+import maybeDeepFreeze from '../util/maybeDeepFreeze';
+
 
 import {
   NetworkStatus,
@@ -195,7 +197,7 @@ export class ObservableQuery<T> extends Observable<ApolloQueryResult<T>> {
     };
 
     return this.queryManager.fetchQuery(this.queryId, combinedOptions, FetchType.refetch)
-      .then(result => this.queryManager.transformResult(result));
+    .then(result => maybeDeepFreeze(result));
   }
 
   public fetchMore(
@@ -277,7 +279,7 @@ export class ObservableQuery<T> extends Observable<ApolloQueryResult<T>> {
         if (options.onError) {
           options.onError(err);
         } else {
-          console.error('Unhandled GraphQL subscription errror', err);
+          console.error('Unhandled GraphQL subscription error', err);
         }
       },
     });
@@ -360,7 +362,7 @@ export class ObservableQuery<T> extends Observable<ApolloQueryResult<T>> {
         ...this.options,
         variables: this.variables,
       } as WatchQueryOptions)
-        .then(result => this.queryManager.transformResult(result));
+      .then(result => maybeDeepFreeze(result));
     }
   }